arrange files, remove useless things
