+++ aircraft_db/.gitignore
@@ -1,0 +1,3 @@
+*.csv.tar
+
+*.csv.gz
