--- dataset/.gitignore
@@ -1,3 +1,0 @@
-*.csv.tar
-
-*.csv.gz
