+++ webapp/.gitignore
@@ -1,5 +1,4 @@
 venv/
-webapp/static/
 
 *.pyc
 __pycache__/
