--- webapp/webapp/templates/base.html
@@ -1,38 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>{% block title %}Mon blog sur les crêpes au sucres{% endblock %}</title>
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.4.1/css/bootstrap.min.css">
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.4.1/jquery.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.16.0/umd/popper.min.js"></script>
-    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.4.1/js/bootstrap.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/Chart.js/2.5.0/Chart.min.js"></script>
-    <script src="https://cdn.jsdelivr.net/npm/chart.js@2.8.0"></script>
-</head>
-<body>
-    <header class="container-fluid h-10 bg-info text-white pl-5 pt-3 pb-3">
-        <h1>AFDA</h1>
-        <p>Un slogan accrocheur</p>
-    </header>
-    <nav class="navbar navbar-expand-sm bg-light">
-
-    <!-- Links -->
-    <ul class="navbar-nav">
-         <li class="nav-item">
-        <a class="nav-link" href="{{ url_for('content.airlines') }} ">Airlines</a>
-        </li>
-        </li>
-        <li class="nav-item">
-        <a class="nav-link" href="{{ url_for('content.compare_airline') }}">Compare Airline</a>
-        </li>
-    </ul>
-
-    </nav>
-
-    <section>{% block section %}{% endblock %}</section>
-
-</body>
-
-<footer style="footer">{% block footer %}{% endblock %}</footer>
-</html>
