--- webapp/webapp/templates/content/airlines.html
@@ -1,37 +1,0 @@
-{% extends 'base.html' %}
-{% block title %}Airlines{% endblock %}
-{% block section %}
-
-<div class="container">
-    <div class="alert alert-info col-3 m-3 float-left text-center">
-        <h2>Number of airlines</h2>
-        <h4>{{ nb_airlines }}</h4>
-    </div>
-
-    <div class="alert alert-info col-3 m-3 float-left text-center">
-        <h2>Number of flights</h2>
-        <h4>{{ nb_flights }}</h4>
-    </div>
-</div>
-
-<div class='container'>
-{{ fig|safe }}
-</div>
-
-
-<div class='container'>
-{% for airline, N in airlines.iterrows() %}
-<div class="card m-3 float-left" style="width: 18rem;">
-    <img class="card-img-top" style="height:2cm;" src="{{ url_for('static', filename=airline+'.png') }}">
-    <div class="card-body">
-        <h5 class="card-title">{{ airline }}</h5>
-        <ul class='list-group list-group-flush m-3'>
-            <li class="list-group-item">Number of flights: {{ N[0] }}</li>
-        </ul>
-        <a href="{{ url_for('content.desc_airline', airline=airline) }}" class="btn btn-primary">See all descriptors</a>
-    </div>
-</div>
-{% endfor %}
-</div>
-
-{% endblock %}
