--- webapp/webapp/templates/content/compare_airline.html
@@ -1,55 +1,0 @@
-{% extends 'base.html' %}
-
-{% block title%}Compare Airline{% endblock %}
-{% block section %}
-<form class='container form-inline' method='POST'>
-    <div class="form-group m-3">
-        <label for="airline1">Airline 1</label>
-        <select class="form-control" id="airline1" name="airline1">
-            {% for airline in airlines %}
-            {% if airline['airline'] == airline1 %}
-            <option selected>{{ airline['airline'] }}</option>
-            {% else %}
-            <option>{{ airline['airline'] }}</option>
-            {% endif %}
-            {% endfor %}
-        </select>
-    </div>
-    <div class="form-group m-3">
-        <label for="airline2">Airline 2:</label>
-        <select class="form-control" id="airline2" name="airline2">
-            {% for airline in airlines %}
-            {% if airline['airline'] == airline2 %}
-            <option selected>{{ airline['airline'] }}</option>
-            {% else %}
-            <option>{{ airline['airline'] }}</option>
-            {% endif %}
-            {% endfor %}
-        </select>
-    </div>
-    <div class="form-group m-3">
-        <label for="airline2">Descriptor:</label>
-        <select class="form-control" id="descriptor" name="descriptor">
-            {% for desc in descriptors %}
-            {% if desc == descriptor %}
-            <option selected>{{ desc }}</option>
-            {% else %}
-            <option>{{ desc }}</option>
-            {% endif %}
-            {% endfor %}
-        </select>
-    </div>
-    <button type="submit" class="btn btn-primary">Submit</button>
-</form>
-
-{% if compute_chart %}
-<div class="container float-left col-8">
-{{ fig|safe }}
-</div>
-
-<div class="alert alert-info col-3 m-3 float-left text-center">
-    <h2>Distance between density</h2>
-    <h4>{{ dist }}</h4>
-</div>
-{% endif %}
-{% endblock %}
